Handle parameter entry with keyword `parameters`
If a parameter entry is mistakenly prefixed with `- parameters`,
previously we'd see it as a multi-parameter header. Before we proceed to
assign the line to that state, take a look at characters before `:` and
after the keyword. Only proceed as parameter header if there's only
whitespace. Otherwise, try parsing the line again as a parameter entry.

Closes #192.